gggg
A Start on the game
