--- COMP3320Project.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.10.35122.118
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "COMP3320Project", "COMP3320Project\COMP3320Project.vcxproj", "{5C57950C-2C3A-40F0-AA7D-84188375EC4A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5C57950C-2C3A-40F0-AA7D-84188375EC4A}.Debug|x64.ActiveCfg = Debug|x64
-		{5C57950C-2C3A-40F0-AA7D-84188375EC4A}.Debug|x64.Build.0 = Debug|x64
-		{5C57950C-2C3A-40F0-AA7D-84188375EC4A}.Debug|x86.ActiveCfg = Debug|Win32
-		{5C57950C-2C3A-40F0-AA7D-84188375EC4A}.Debug|x86.Build.0 = Debug|Win32
-		{5C57950C-2C3A-40F0-AA7D-84188375EC4A}.Release|x64.ActiveCfg = Release|x64
-		{5C57950C-2C3A-40F0-AA7D-84188375EC4A}.Release|x64.Build.0 = Release|x64
-		{5C57950C-2C3A-40F0-AA7D-84188375EC4A}.Release|x86.ActiveCfg = Release|Win32
-		{5C57950C-2C3A-40F0-AA7D-84188375EC4A}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {3DF20BC5-9C45-42D4-BB96-7A24D01F304C}
-	EndGlobalSection
-EndGlobal
