+++ COMP3320Project/COMP3320Project/COMP3320Project.vcxproj.filters
@@ -15,7 +15,7 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="scr\Application.cpp">
+    <ClCompile Include="scr\main.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
