+++ COMP3320Project/COMP3320Project/x64/Debug/COMP3320Project.log
@@ -1,2 +1,5 @@
-﻿  Application.cpp
+﻿  main.cpp
+glew32s.lib(glew.obj) : warning LNK4099: PDB '' was not found with 'glew32s.lib(glew.obj)' or at ''; linking object as if no debug info
+LINK : warning LNK4098: defaultlib 'MSVCRT' conflicts with use of other libs; use /NODEFAULTLIB:library
+LINK : warning LNK4098: defaultlib 'LIBCMT' conflicts with use of other libs; use /NODEFAULTLIB:library
   COMP3320Project.vcxproj -> C:\Users\Alex\dev\2024\project\COMP3320Project\x64\Debug\COMP3320Project.exe
