+++ COMP3320Project/COMP3320Project/x64/Debug/COMP3320Project.tlog/Cl.items.tlog
@@ -1,1 +1,1 @@
-C:\Users\Alex\dev\2024\project\COMP3320Project\COMP3320Project\scr\Application.cpp;C:\Users\Alex\dev\2024\project\COMP3320Project\COMP3320Project\x64\Debug\Application.obj
+C:\Users\Alex\dev\2024\project\COMP3320Project\COMP3320Project\scr\main.cpp;C:\Users\Alex\dev\2024\project\COMP3320Project\COMP3320Project\x64\Debug\main.obj
