+++ COMP3320Project/COMP3320Project/x64/Debug/COMP3320Project.tlog/link.secondary.1.tlog
@@ -1,2 +1,2 @@
-^C:\USERS\ALEX\DEV\2024\PROJECT\COMP3320PROJECT\COMP3320PROJECT\X64\DEBUG\APPLICATION.OBJ
+^C:\USERS\ALEX\DEV\2024\PROJECT\COMP3320PROJECT\COMP3320PROJECT\X64\DEBUG\MAIN.OBJ
 C:\Users\Alex\dev\2024\project\COMP3320Project\COMP3320Project\x64\Debug\COMP3320Project.ilk
